+++ Assets/Scripts/Scenes/InGame/System/EscapeItem.cs
@@ -2,13 +2,11 @@ using UnityEngine;
 using UniRx;
 using Cysharp.Threading.Tasks;
 using System.Threading;
-
-namespace Scenes.Ingame.Manager
+using Scenes.Ingame.Manager;
+namespace Scenes.Ingame.InGameSystem
 {
-    public class EscapePoint : MonoBehaviour
+    public class EscapeItem : MonoBehaviour
     {
-        [SerializeField]
-        private Material _activeMaterial;
         IngameManager manager;
         private bool _isActive = false;
         private bool _get = false;
@@ -17,23 +15,24 @@ namespace Scenes.Ingame.Manager
         {
             token = new CancellationTokenSource();
             manager = IngameManager.Instance;
-            manager.OnInitial.Subscribe(_ => _isActive = false).AddTo(this);
-            manager.OnOpenEscapePointEvent.Subscribe(_ =>
-            {
+            manager.OnInitial.First().Subscribe(_ => _isActive = false).AddTo(this);
+            manager.OnIngame.First().Subscribe(_ =>
+            { 
                 _isActive = true;
-                GetComponent<Renderer>().material = _activeMaterial;
                 GetItem(token.Token).Forget();
             }).AddTo(this);
+
         }
         async UniTaskVoid GetItem(CancellationToken token)
         {
             await UniTask.WaitUntil(() => _get);
-            manager.Escape();
+            manager.GetEscapeItem();
             Destroy(gameObject, 0.5f);
         }
         private void OnTriggerEnter(Collider collision)
         {
             if (_isActive == false) return;
+            if (collision.gameObject.CompareTag("Player"))
             {
                 _get = true;
             }
