+++ Assets/Scripts/Scenes/InGame/System/ReadyEnum.cs
@@ -1,7 +1,7 @@
 /// <summary>
 /// IngameManageの_initialEvent発行後の_ingameEventが発行されるまでに必要な準備作業
 /// </summary>
-namespace Scenes.Ingame.Manager
+namespace Scenes.Ingame.InGameSystem
 {
     public enum ReadyEnum
     {
