+++ Assets/Scripts/Scenes/InGame/System/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f8c0e9440e509f5459b97622ab4ad5dd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
